requirements.txt
Django==1.6.5
Markdown==2.4.1
Pillow==2.4.0
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-allauth==0.16.1
django-markdown==0.6.1
django-toolbelt==0.0.1
gunicorn==18.0
oauthlib==0.6.3
psycopg2==2.5.3
pystache==0.5.4
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.2
static==1.0.2
